feat: refactor import path for ExperimentVisitsTable and add new component for visit registrations
